requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.3
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.0
MarkupSafe==2.0.0
packaging==20.9
Pillow==8.2.0
psycopg2==2.8.6
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
